Added Eric Gudgion's feedback
In the objective, defined the MOJO SP and Python SP.
In the prerequisites, added a comment for non-aws lambda deployment, so users know we
will also do MOJO REST Server Deployment.
After figure 2, added Eric's suggested content on if a user needs to do batch scoring
in production, they should use a standalone scorer instead of Score on Another Dataset
feature.
In How To Choose a Scoring Pipeline Deployment, added a column on scoring pipeline server deployment and more content to the rest of the table.
At the start of Task 3, added a reminder that we use Score on Another Dataset feature
in the tutorial, but in production, one should use a Driverless AI Deployment Template.